moved child structs down to channel
