+++ lib/Structs/Channels/Invite/invite_channel.ex
@@ -1,24 +1,7 @@
 defmodule Alchemy.Channel.Invite.InviteChannel do
-  @moduledoc """
-  Represents the channel an invite is for
+  @moduledoc false
 
-  - `id`
 
-    The id of the channel
-  - `name`
-
-    The name of the channel
-  - `type`
-
-    the type of the channel, either "text" or "voice"
-  """
-  @type snowflake :: String.t
-
-  @type t :: %__MODULE__{
-    id: snowflake,
-    name: String.t,
-    type: String.t
-  }
   @derive Poison.Encoder
   defstruct [:id,
              :name,
