+++ lib/Structs/Channels/channel.ex
@@ -1,5 +1,8 @@
 defmodule Alchemy.Channel do
   alias Alchemy.OverWrite
+  alias Alchemy.DMChannel
+  alias Alchemy.Channel.{Invite, Invite.InviteChannel, Invite.InviteGuild}
+  import Alchemy.DMChannel, only: [channel_type: 1]
   import Alchemy.Structs.Utility
   @moduledoc """
   This module contains useful functions for operating on `Channels`.
@@ -24,9 +27,6 @@ defmodule Alchemy.Channel do
   - `position`
 
     sorting position of the channel
-  - `is_private`
-
-    should be false for guild channels
   - `permission_overwrites`
 
     an array of `%OverWrite{}` objects
@@ -56,6 +56,26 @@ defmodule Alchemy.Channel do
     user_limit: Integer | nil
   }
   @typedoc """
+  DMChannels represent a private message between 2 users; in this case,
+  between a client and a user
+
+  - `id`
+
+    represents the private channel's id
+  - `recipient`
+
+    the user with which the private channel is open
+  - `last_message_id`
+
+    the id of the last message sent
+  """
+  @type dm_channel :: %DMChannel{
+    id: String.t,
+    type: atom,
+    recipients: User.t,
+    last_message_id: String.t
+  }
+  @typedoc """
   Represents a permission OverWrite object
 
   - `id`
@@ -77,6 +97,98 @@ defmodule Alchemy.Channel do
     allow: Integer,
     deny: Integer
   }
+
+  @type snowflake :: String.t
+  @type hash :: String.t
+  @type datetime :: String.t
+
+  @typedoc """
+  Represents an Invite object along with the metadata.
+
+  - `code`
+
+    The unique invite code
+  - `guild`
+
+    The guild this invite is for
+  - `channel`
+
+    The channel this invite is for
+  - `inviter`
+
+    The user who created the invite
+  - `uses`
+
+    The amount of time this invite has been used
+  - `max_uses`
+
+    The max number of times this invite can be used
+  - `max_age`
+
+    The duration (seconds) after which the invite will expire
+  - `temporary`
+
+    Whether this invite grants temporary membership
+  - `created_at`
+
+    When this invite was created
+  - `revoked`
+
+    Whether this invite was revoked
+  """
+  @type invite :: %Invite{
+    code: String.t,
+    guild: invite_guild,
+    channel: invite_channel,
+    inviter: User.t,
+    uses: Integer,
+    max_uses: Integer,
+    max_age: Integer,
+    temporary: Boolean,
+    created_at: datetime,
+    revoked: Boolean
+  }
+  @typedoc """
+  Represents the guild an invite is for.
+
+  - `id`
+
+    The id of the guild
+  - `name`
+
+    The name of the guild
+  - `splash`
+
+    The hash of the guild splash (or nil)
+  - `icon`
+
+    The hash of the guild icon (or nil)
+  """
+
+  @type invite_guild ::%InviteGuild{
+    id: snowflake,
+    name: String.t,
+    splash: hash,
+    icon: hash
+  }
+  @typedoc """
+  Represents the channel an invite is for
+
+  - `id`
+
+    The id of the channel
+  - `name`
+
+    The name of the channel
+  - `type`
+
+    the type of the channel, either "text" or "voice"
+  """
+  @type invite_channel :: %InviteChannel{
+    id: snowflake,
+    name: String.t,
+    type: String.t
+  }
   @derive Poison.Encoder
   defstruct [:id,
              :guild_id,
@@ -89,16 +201,6 @@ defmodule Alchemy.Channel do
              :bitrate,
              :user_limit]
 
-
-  @doc false
-  def channel_type(code) do
-    case code do
-      0 -> :text
-      1 -> :private
-      2 -> :voice
-      3 -> :group
-    end
-  end
 
 
   @doc false
